HTML Setup: Add DB name and server port as hidden fields in wizard.
Add the DB name and DB server port as hidden fields so that we can
use these when testing the DB connection.  If the user as changed
one or both of these in the advanced setup screen, the test could
fail unless we use these saved values in the wizard screen.