HTML Setup: Storage Group Editor changes.
Allow creating new Storage Groups.

Reload the SG list to remove a SG tab when the last directory in that
SG is deleted.

Remove the 'Delete Storage Group' button placeholder.  It is almost
just as easy to click the 'Delete' button on each directory and less
code, now that the SG tab goes away when the last directory is removed.