Handle jump points through the frontend http interface.
This does not attempt to filter the jump on the basis of the current
location (e.g. actioning 'Live TV' when in live tv will restart live
tv). Filtering requires a key to be bound to the jump point (most, if
not all, default to unbound) but requiring a key binding does not make
sense in the context of directly triggering actions from an external
source and is inconsistent with the handling of all other actions via
the http interface (which do not require a keybinding).

The network control code attempts to test whether a jump has succeeded
by testing the result of GetCurrentLocation but, as commented in that
code, the results are completely unreliable.