HTML Setup: Javascript and tabs simplification and bugfix
Move tab initialization to a function with tab preloading
in the background turned on by default.

Disable tab animation for the Storage Groups Editor since it breaks
the 'add new storage group' functionality due to what appears to be
a bug when we add a new tab then select it.  The new tab is not
selected when we try to select it.

Remove some unused javascript functions for adding/removing
tabs since they are only wrappers for 1-line jquery calls.

Rename 'setuptabs' on the JobQueue editor to 'jobqueuetabs'.

Init the Storage Group Editor tabs only once, there's no need to
reinit every time we initStorageGroups() since we call that
function during several operations now.