API: Add a SendMessage method to the Myth service.
More or less exactly approximates the functionality of the mythmessage program, but is not usable from a browser, scriptable, callable from JSON, etc.  Causes a message to appear on all frontends, or on a single frontend.

Usage:

http://BackendServerIP:6544/Myth/SendMessage?Message=Here is some message text&Address=192.168.0.5&udpPort=6948

Address and udpPort arguments are both optional.  They default to 255.255.255.255 (all connected frontends) and 6948, respectively.  So send a message to all frontends at once, quickly:

http://BackendServerIP:6544/Myth/SendMessage?Message=The system is going down!