HTML Setup/Backend Web Server: Add message sending capabilities.
Adds a button/page to send messages to frontends, defaulting to broadcast to all frontends.

There are a couple of improvements I'd like to see here, such as making the list of frontends a dropdown instead of text entry and validating the UDP port value... but this works pretty nicely for now.

Adds a util function to show and hide a div that might come in handy later on.  Here it's used to hide advanced options.