TV: Create an Advanced menu option in the OSD Video menu.
Create the new menu item, move the video scan selection into the new
menu (most users will never need to touch this) and create a new
additional sub-menu to override the current deinterlacer.

The deinterlacer selection is disabled for the time being until
MythPlayer is re-factored a little to better deal with customising the
deinterlacer. Currently, selecting a new deinterlacer will not change,
for example, whether the player runs at double rate or not.