Only check for an updated theme version when at the main menu.
Instead of pinging the MBE every hour to check for a new version of
the currently running theme, only do so if we're at the main menu
when the new theme notification would actually be displayed.  This
will probably get more tweaking in the future, so that we have a
better chance of notifying the user rather than the current method
of firing a timer every hour to run the check routine.