HTML Setup: Don't display extra headers.
Do date, we were displaying a content page title ("Database Settings") and a tab name ("General") and then a third redundant title in the tab, generated by the backend mythsettings class.  Rather than remove it from the server, simply make it invisible (as the information has value, and maybe someone will want to write a setup skin to show it-- but it's not great for our layout).