HTML Setup: Add a confirmation dialog.
Use the JQuery dialog qidget to build a basic confirmation dialog.  Arguments are the message, a function name, and an optional array of parameters.

So, to perform javascript function foo with params bar when the answer is "yes,":

showConfirm("Are you ready to perform function foo?", foo, bar);

To perform javascript function foo, which takes no parameters:

showConfirm("Are you ready to perform function foo?", foo);