Handle decoded frame number of -1 more gracefully
Fixes #9699.  The proposed fix in the ticket has been implemented.  As the
frameNumber can be -1, the type of lastDecodedFrameNumber must be signed.  This
is to fix a premature exit during transcoding, but it quite likely could rear
its ugly head in other contexts too.

A better permanent fix would be to treat the special case of frameNumber == -1
and deal with it specifically.

Thanks to ghaushe for the find.