HTML Setup: Convert 'Add new SG Dir' & 'Add new SG' to #edit dialogs
This converts the 'Add new SG Dir' to use the existing 'edit' div
dialog and also converts the 'Add new SG' functionality to use the
edit dialog as well.

Comment out the 'Create new Storage Group' button in favor of a '+' tab
in the Storage Groups editor.

Fix a bug in addStorageGroupDir() that I created in my previous alert()
cleanup.