HTML Setup: Status/Error message bugfix and SG validity checking
If a new status or error message comes in while the message is
still being displayed, clear the close timer for the previous
message and start a new one so that the new message displays for
the desired duration.

Perform some validity checking when we save a new Storage Group
directory.