API: Add a new "Channel" service.
Intention is to expand this to provide the many, many methods needed to perform channel/lineup/etc. editing, but at the moment the utility is somewhat limited by the need to expand the framework to pass through certain data types that it doesn't understand yet (and which I haven't figured out how to teach it).

Still, there are a couple of little things in here to start (and if someone wants to make the API pass through/render QList<uint> so we can get a list of ChanIDs I won't complain).