API: Rework the Channel service.
Provide the ability to get a list of channels, with associated basic information, for a given multiplex.  There's still some more work needed here.  Need to make sure we are accurately filling in some of these values, some of which I am not sure how to get without resorting to SQL.  Also need to reconcile what's needed to edit a channel and make sure we are providing it with the channelinfo object.  Finally, need to add the ability to pass a channel back to the service/modify it.