API: Additions to the Channel service.
Add a number of other details about a channel which we will need to edit them, though a few more are still missing.  Also switch to using GetChannelData frmo channelutil when possible as it's a more efficient way of getting most of this information.

I'll need to expand channelutil a bit to get a few more of the details such as visibility, etc.