Fix video buffering code to properly play high reference frame videos.
This change prevents a buffer to be returned to the available queue while it's still in use by the decoder. There is now a 'finished' queue to store frames until the decoder no longer needs them. Previously we would return a frame to available before it was finished by the decoder and cause severe video corruption.

With this change, killasample can now be played corruption free by Xv and by VDPAU with vdpaubuffersize of only 20.

NOTE: Bringing up the OSD under Xv can still cause corruption since we don't protect against modifying the buffer in the video processing portion of code. To avoid this the buffer will need to be copied before being modified.