HTML Setup: Add the beginnings of a channel editor.
Extreme work in progress.  You can't select the sourceid yet, so if you want to use a sourceid besides '3,' you need to set the sourceid in the channeleditor.html file for now.  Also, no editing just yet. this is just a channel browser at the moment.

With all of that said, all the necessary fields to edit a channel are present (with most of them hidden), so changing this to a true channel editor just means writing a dialog with the user-facing options, and calling the API method to update channels on return.

I need to add a method to return a map of video sources and names so that you can switch between sources.  Will try to get this done later today.