API/HTML Setup: Various Channel Editor Fixes.
Cast frequency to a long since the API doesn't understand how to pass through uint64_t, but long should do the job.  I found in the backend logs that this was causing all the slowdown in rendering the channel list, now it's much quicker.

Reorganize some of the channel editor javascript functions.