VideoOutpuut: Add 2 new video fill modes.
These fill the screen either horizontally or vertically with cropping
top and bottom or left and right respectively.

These will be no-ops when no adjustment is required and are useful for
making full use of all the available screen space.

The existing fill stretch modes cater to incorrectly encoded video (and
I believe should be removed, as they should be handled with a
combination of aspect and fill adjustment) and the half and full fill
modes are 'hard coded', both in terms of the amount of fill applied and
when it is applied (i.e. all of the time).

Closes #9035.