HTML Setup: Change the way settings are formatted by the backend.
Use the html <label> tag in line with best practices for location of the label.

Label goes directly after the checkbox, and the label tags enclose the widget itself.  Labels can also be clicked to activate the widget for better use on touch devices.

Label goes before the widget for input and select widgets.

Use CSS to increase the padding between settings.

Put the help links unconditionally at the end of the line.