HTML Setup: Introduce the idea of simultaneous multichannel editing.
(the following actions are not hooked up to actual saves yet)

This (currently very simple) wizard applies an action to an array of channels, using templated user input.

So, you would select a group of channels in the editor, and hit the edit button or right click->edit.  You would select which field you want to rename, construct your templated text using the provided box of options, like so:

http://www.fecitfacta.com/multiedit.png

And that rename action would be applied to all selected channels.  This wizard could be expanded such that the verb is a select box as well, ie rename, hide, show, etc.

Mostly just an exercise to see how the concept feels, but I think it's worth carrying through to completion.