Channel Icons in Storage Groups
Make PixmapChannel::CacheChannelIcon() look in the 'ChannelIcons'
virtual Storage Group on the MBE rather than in the 'Default'
Storage Group for channel icons.  If for some reason, users have
channel icons in their Default Storage Group, then they should
continue to work correctly because the Storage Group code does
a hunt of other Storage Groups if it can't find the file in the
given group.