This commit implements a new command line parser, extensible using a QVariantMap of responses, rather than pre-defined class attributes.
Usable options are added using a series of overloaded 'add()' methods,
following the format:
    add(<command line option>, <stored name>, <default>,
        <short help>, <long help>)

The command line option can be either a single string, or a QStringList
containing multiple aliases to be used for the option. The stored name
is what will be used to retrieve the value after processing. The add()
methods currently support the following types of defaults:
    bool            int             uint
    long long       double          const char *
    QString         QSize           QDateTime

Other types will have to take the general format:
    add(<command line option>, <stored name>, QVariant::Type,
        QVariant<default>, <short help>, <long help>)

Several custom classes have been made pre-built with the needed
arguments for their respective binary:
    MythBackendCommandLineParser
    MythFrontendCommandLineParser
    MythPreviewGeneratorCommandLineParser
    MythWelcomeCommandLineParser
    MythAVTestCommandLineParser
    MythCommFlagCommandLineParser
    MythJobQueueCommandLineParser
Additional classes are defined in the header, but have not yet been
configured. The above were chosen because those binaries already used
the pre-existing MythCommandLineParser.

Once created, all necessary command line processing is done internal to
the class, with no need for an external loop. Since new options can be
added dynamically, there is no longer a need for external access to this
loop to provide support for options not already included in
MythCommandLineParser. The internal processing loop is called using:
    MythCommandLineParser cmdline;
    if (!cmdline.Parse(argc, argv))
        return GENERIC_EXIT_NOT_OK;

Boolean options do not expect an associated value, and will be set to
true, or if given a default, the opposite of their default, when
received.  QString options can optionally have an associated value, and
if not, will be stored as using their given default value.  All other
option types are required to have a value associated. Values must not
have spaces, and must not start with '-'. A dash will trigger a fault as
an invalid option.

QSize options are required to be in the format of '<X>x<Y>', or will
trigger a fault. QMap options are required to be in the format of
'<key>=<value>', and are allowed to be called multiple times, such as
used with the --override-settings option. Additional values will be
collected as independent arguments, and must come after all defined
options. An argument received before an option will trigger a fault.

Once parsed, --override-settings are available as before through
GetSettingsOverride, and additional arguments are available as a
QStringList through GetArgs. Options are available through the '[]' map
operator, or a series of 'to<Type>()' methods, using the stored name.
The following type conversions are available:
    bool                    toBool()
    int                     toInt()
    uint                    toUInt()
    long long               toLongLong()
    double                  toDouble()
    QSize                   toSize()
    QString                 toString()
    QStringList             toStringList()
    QMap<QString, QString>  toMap()
    QDateTime               toDateTime()

For boolean options, toBool() will return the value if set, or else the
default. For all other types, toBool() will return whether or not that
option was set on the command line. All other to<Type> methods will
return the given value if available, or fall back to the set default.

Two other utility methods are available. PrintVersion() will print the
version information as before. PrintHelp() will generate help information
for all registered options. If '--help' is called without a value,
PrintHelp() will print all available options, along with their short
help string. If the short help string is left blank, the options will be
left hidden, and not displayed by this general printout. If '--help' is
called with a value, then a more in-depth page is printed, with aliases
of that option, type, default value, and the long help. If the long help
is left blank, the short help string is used in its place. For both the
short and long help, newline characters are honored, and the entire
block is indented as needed.