HTML Setup: Fixes and a channel icon browser
Make the JobQueue job editor popup dialog auto-size vertically.

Allow highlighting files in the file browser.

Fix the Storage Group editor following recent table header conversion
from tr to thead.

Hook up a simple channel icon browser for now.  This currently just lists
channel icons stored in the ChannelIcons Storage Group on the MBE.  The
default dir for this virtual Storage Group is MYTHCONFDIR/ChannelIcons
(ie, ~/.mythtv/ChannelIcons on a *nix machine).  This is just a simple
browser for existing icons, the plan is to integrate downloading of remote
icons as well.  A preview window will also be added later.

Rework the file browser to allow it to be opened in one of three ways:
 1) Directory browsing/selection mode
 2) File browsing/selection mode
 3) Storage Group file browsing/selection mode