Change the AutoExpire UpdateThread to not self-delete
It seems that using this->deleteLater() as the last line in the thread's
execution can still cause an immediate deletion by the event thread, even before
the victim thread has finished executing.  This causes a segfault.  So, instead,
for detached threads whose pointers we keep, we can check that they are running
before starting a new one.

This has fixed a repeatable crashing with GreyFoxx's setup, and hopefully will
stabilize for others as well.