Add support for a single '-' as a string value, for use when defining stdin/stdout.
Add options for mythfilldatabase into MythFillDatabaseCommandLineParser
and migrate mythfilldatabase over to using it. Due to limitations of the
class, there are a few changes in the format.

The --file, --ddfile, and --xawchannels options are now boolean and no
longer accept values.  They are now used to set the operational mode,
and will not function without other arguments being supplied.  These
options now require the following additional options.

--file
    --sourceid <int>
    --xmlfile <string>

--ddfile
    --sourceid <int>
    --offset <int>
    --lineupid <int>
    --xmlfile <string>

--xawchannels
    --sourceid <int>
    --xawtvrcfile <string>