Rework the handling of missed recordings to be more efficient.
Intead of deleting and rewriting all future entries every time the
scheduler runs, only delete entries that are no longer applicable and
only write entries that changed.  This also should remove the
possibility of losing entries if the backend crashes or hangs while
the scheduler is running.

In addition, this change cleans up the logic a little and fixes some
minor issues with the previous change.  Finally, it also fixes a
known, longstanding, bug where the schedTime variable had file static
scope and could possibly cause a problem if a speculative scheduleer
was running at the same time as the main scheduler.