Sort recordings for bindings and MythArchive.
This change causes ProgramInfo::LoadFromRecorded() to always sort
recordings according to starttime (descending) in 0.24-fixes (only).  A
better fix was committed to master, but due to the binary version and
protocol version changes, can't be backported to 0.24-fixes.  This
approach will make the list more usable for MythArchive and bindings
users, but without any version changes.

Refs #9486.