Backend Settings/HTML Setup: change layout and implement placeholder text.
Re-engineer the layout of the backend-served settings a little to follow some more contemporary form layout.  This still all needs work but I think it's headed in the right direction.

Implement placeholder_text parsing in the backend settings xml parsing.  This takes advantage of a feature in modern browsers to offer a hint as to what goes in a given input box.  Adjust the existing settings to use placeholder text.

Change the CSS a bit to put some padding below each setting to make visual tracking a little easier.  I took a quick survery of the form layout approach followed by 10 or so major websites, and this same sort of widget and tag usage seemed fairly prevalent/commonplace, so if it's good enough for them...