New asynchronous mythlogging support.
All logging is done via a queue and a thread that will send the logs to several
log outputs (including console, debug log file, syslog, database).  There is no
blocking the thread that is sending the log message, and the log messages are
QString-free (as QString is not thread-safe).

As logging is asynchronous, it is quite possible that issues that formerly were
hidden or helped by the blocking in the old logging will surface.  However, the
threads should work as expected.

On systems that have gettimeofday, you get microsecond accuracy on the logs.

Threads (and QRunnables) need to register and deregister themselves with the
logging if the thread name is to appear in the logs rather than
"thread_unknown".

Still need to flesh out DatabaseLogging.  By default will log to a logfile if
provided, the console and to LOCAL7 in syslog.