libmythui: Stop the main window background from automatically painting.
This sets 2 properties on the MythMainWindow widget (N.B. not the
painter widget) to prevent Qt from continually repainting the theme
background.

Mythfrontend startup is now noticably faster and the brief flash of the
theme background when starting video playback is gone, giving a slight
speed improvement there as well.

I can't see any adverse effects on the several frontends tested with a
variety of different video drivers and render options.