Fix the check to see if we have a valid DB schema.
Change the code that checks to see if we have a schema available so that
it doesn't use a count of the value returned by QSqlDatabase::tables().
The tables() function returns a list of all tables visible to the
user--even tables in other databases--so can result in false positives.

Fixes #9742.
(cherry picked from commit eda0880f93f6de4c501ae7304787825fc5cdd588)