HTML Setup: Add a "Last 10 Recorded" widget to the overview page.
Just a fun little addition that makes the overview page more interesting and fun.  I have a few other ideas for worthwhile widgets for the top page, but this is a nice start and should be fun for those looking at the backend web server for the first time (at least, those with some recordings).

Parts shamelessly cribbed from the GetRecorded example.