API: Add a Video Service.
Adds a service for Video metadata (since videometadata is now a DB table independent of a plugin, and MythVideo as a plugin will soon cease to exist).

Right now, just returns a list of (unfortunately named to prevent class name collisions with libmythmetadata) VideoMetadataInfo objects using API calls nearly identical to the GetRecordings API.

Use:

http://BackendServerIP:6544/Video/GetVideos?StartIndex=10&Count=25

To return 25 video metadata items, starting at the tenth.

While I was add it, I added methods to the Content Service to return Fanart, Banners, and Screenshots for Video, and renamed the "GetVideoArt" API to match its correct naming elsewhere in our code for that art type, GetVideoCoverart.  I need to go back in and extend these functions to resize as the other artwork calls do.