Fix MHEG-related crash post ffmpeg-sync
Two changes in here:

One: don't try to dereference ->info if it is NULL
Two: use av_new_stream instead of av_add_stream.  The contents are nearly
     identical, but the former actually allocates the ->info.

I don't think the ->info part is actually used by MHEG itself, but enough of
the generic code for mpeg2ts seems to use it that this seemed to be the best
way to fix this crash.

Fixes #9747