API: Add a Blu-ray metadata parser.
Limited use API alert!  Honestly not even sure what I'll use it for, but could be of substantial use to those seeking to write remote control applications.

Can be used on a specific path or the mounted disc.  Returns all the Metadata for the Blu-ray title-- if the Disc also has DiscLibrary Metadata such as thumbnail, etc. it returns the path to the thumbnail, disc titles, language, etc.

Takes one argument, Path.  If you don't supply a path, then it defaults to your Blu-ray mount point and attempts to parse the disc in the drive for info.  So:

http://BackendServerIP:6544/video/GetBluray

(will return info for the mounted disc)

http://BackendServerIP:6544/video/GetBluray?Path=/path/to/a/folder/with/BDMV/in/it/

(will return info for a BD folder structure at that path)

Modifies the library API, distclean and recompile plugins.