API: Add an AddVideo method to the Video Service.
Usage (POST-only):

http://BackendServerIP:6544/Video/AddVideo?Filename=/path/to/a/file.ext&Host=hostname

Both the Filename and Host arguments are required.  As local video storage is deprecated, it is unsupported in the API and will remain that way.  This code won't work with local storage.

Usual filename parsing is performed on added videos, so Title/Subtitle/Season/Episode will be parsed out as normal from the filename itself.

No metadata is done on this data yet.  The metadata classes are asynchronous and rely on listening for Qt Events, and I can't figure out how to make either work with the API classes yet.