StorageGroup::GetDirFileList() fix to a bug in a bugfix.
Fix the fix to 7255d463 in af458b3ac.  We only need to append the
relative path if we are running in recursive mode.  The only
code that should be using recursive mode currently is the
/Content/GetFileList API call.  This fix to a fix still works with
BD over Storage Groups which is what the original fix fixed.  There,
have I said fix or fixed enough?  If not, I can fix it.