EIT fixups for Kabel Deutschland.
Adds a new transponder and applies EIT fixups to a couple existing ones.

There was some concern about the HDTV fixup being applied to a channel which shows some upscaled content. My response to that, upon consideration, is that:

a) If you want more granular information, you need a real source of listings which doesn't rely on fixups.
b) kingofsat and lyngsat both refer to this transponder as a blanket "HD," so I'm following their lead.