Add support for <i18n> translation tag in .qsp and .qjs files.
Text between <i18n> and </i18n> will be run through QObject::tr()
before being processed by the scripting engine.  The values are
translated as the file is read in, so the translated values are
cached as part of the parsed script.

Translations are not yet enabled in mythbackend, and the HTML
setup webpages do not use the <i18n> tags yet, so there are not
yet any visible effects from this patch.

This modifies the binary API version, so make clean, etc..