HTML Setup: front page speedup, channel icon downloading.
Reorder the front page loading a little bit to give a perceived speed
increase.

Add a 'search' button to the channel editor to query the services.mythtv.org
channel icon database.  Found icons are listed in a grid and an icon may
be downloaded and installed by clicking on the preview.  The preview image
on the main channel detail editor screen still has an aspect ratio issue,
but it is displayed.  This is a manual downloader, we hope to automatically
select the icon for the user, but this manual override may still exist after
that.