Make shutdown smoother by using logStop()
In the MythContext dtor, run logStop() to stop the logging thread to keep from
getting an annoying crash saying a thread was deleted while running.  Also,
tweak the stay-alive time for the ProgramInfoUpdater QRunnable for faster
shutdown, and tell the logging thread to empty its queue on shutdown so all
the log messages are seen.