HTML Setup: Minor layout and behavior tweaks for channel detail editor.
Display the icon buttons inline so that they take up a little less vertical space.

Only display the icon element if there's an icon there.

Don't reload the entire grid after the save of a channel, this isn't necessary and caused a weird reordering of channels if you edited something on the first page before sorting.