Change --logfile to --logpath again
Rework of an earlier patch from Michael T. Dean which had to be removed for
the merge with the new commandline parser.  At this point, a directory can be
handed to the -l commandline option, which will cause a logfile named
application.pid.log to be used.  If -l is used with no argument, the debug log
file will be disabled.