Add a MEDIA menu item type to the main menus.
This type allows you to add menu items that can call one of the media handlers.
For example you could add a menu item to show a particular web page or to play
a video using the internal player.

An example item to run MythBrowser

    <button>
        <type>WEBPAGE</type>
        <text>WebMail</text>
        <description>Show email</description>
        <action>MEDIA WebBrowser https://example.domain.com/signin/webmail/</action>
        <depends>mythbrowser</depends>
    </button>

An example to play a video file

    <button>
        <type>VIDEO</type>
        <text>Show Example Video</text>
        <description>Show an example video</description>
        <action>MEDIA Internal /media/videos/example.mpg/</action>
    </button>