MythMainWindow: Break out the actual UI drawing into a separate method.
If the rendering context is shared, MythMainWindow assumes the 'other'
owner will call the draw method as appropriate (though UI animation will
continue to run as before).

In VideoOutputOpenGL, if we are embedding and the context is shared,
render the main UI after the framebuffer has been cleared and before the
video.