Convert DeleteThread and TruncateThread to QRunnable
Fixes #9757.

It seems that it might be the DeleteThread exiting poorly causing this issue.
As I'd earlier found that the this->deleteLater() code structure is nasty in
QThreads, this was ripe for a redo anyways.  I have changed it to using a
QRunnable (kinda a lightweight QThread) which actually will run in a thread
from a global threadpool, and will clean up after itself.  Since the
TruncateThread was nearly identical in structure, I changed it too.