Fix displaying the frontend settings screens.
With certain combinations of hardware, driver and platform, the frontend
settings pages are not displayed when using the OpenGL painter. This is
because the libmythui painter renders into the painter window whereas
the settings pages (non-libmythui) are still rendered into the
MythMainWindow object (the parent window).

In the majority of cases this works without issue but sometimes the
painter window obscures the parent. As a workaround simply hide the
painter window when entering the settings pages.

A similar workaround may be necessary for any remaining non-libmythui
plugins.