Add ASS/SSA subtitle support.
Based on the final patch from Shawn King (thank you) in #9294 with the
following changes:-

- fixed the display area to line up with the video rect.
- only initialise libass when required.
- various additional NULL pointer checks and minor refactors.
- use SubtitleScreen::m_safeArea in the ass code.

I chose to use a compile time dependency on libass rather than create
another internal library version (and avoid the maintenance that it
would need) or use dynamic linking (extra complexity). As noted in
configure, the libass version (although not actually checked) needs to
be at least 0.9.10 - which is probably still too new for the libass
version packaged with most distros.

Closes #9294