TV: Add Jump ahead and jump back options in the OSD navigate menu.
These are only shown if the program is not a dvd or bluray and there is
no chapter information - in which case the chapter navigation is
available.

The DVD Chapter menu option probably needs removing as it very rarely
actually works and should be replaced with some dvd chapter handling
code.